Add GUI functionality to duke.
Previously, duke was a purely CLI program which interacted
with user through text.

With the new GUI functionality, users are now able to view a GUI
for duke which will help enhance usability and experience for them.

In the future, there is a hope to spiff up the look of the duke
GUI to make it more appealing to users.