Adds better exception handling capabilities to Duke
With this commit, duke is now able to handle more types of events
that can cause the program to not function as expected. An additional
exception is if the stored data was somehow corrupted and its data
is not retrievable. Also, refactored out some code exception handling
code to better reflect locality of exception catching.